feat: Move UI logic from MainActivity to ConnectToServerActivity
Moved the UI logic from MainActivity to ConnectToServerActivity to better organize the codebase. Updated the ConnectToServerView class to reflect these changes.

Files modified:
- Modified MainActivity.java
- Modified ConnectToServerActivity.java
- Modified ConnectToServerView.java
