+++ app/src/main/java/com/example/munchkin/view/ConnectToServerView.java
@@ -4,34 +4,34 @@ import android.widget.Button;
 import android.widget.EditText;
 import android.widget.TextView;
 
-import com.example.munchkin.MainActivity;
+import com.example.munchkin.ConnectToServerActivity;
 import com.example.munchkin.R;
 
 public class ConnectToServerView {
 
-    private MainActivity mainActivity;
+    private ConnectToServerActivity connectToServerActivity;
     private TextView textViewServerResponse;
 
     private EditText editTextUsername;
 
-    public ConnectToServerView(MainActivity mainActivity) {
-        this.mainActivity = mainActivity;
-        this.textViewServerResponse = mainActivity.findViewById(R.id.responseTextView);
-        this.editTextUsername = mainActivity.findViewById(R.id.usernameEditText);
+    public ConnectToServerView(ConnectToServerActivity connectToServerActivity) {
+        this.connectToServerActivity = connectToServerActivity;
+        this.textViewServerResponse = connectToServerActivity.findViewById(R.id.responseTextView);
+        this.editTextUsername = connectToServerActivity.findViewById(R.id.usernameEditText);
         setupUI();
 
     }
 
     private void setupUI() {
-        Button buttonSendMsg = mainActivity.findViewById(R.id.sendbutton);
+        Button buttonSendMsg = connectToServerActivity.findViewById(R.id.sendbutton);
         buttonSendMsg.setOnClickListener(v -> {
             String username = editTextUsername.getText().toString();
-            mainActivity.sendMessage(username);
+            connectToServerActivity.sendMessage(username);
         });
 
 
-        Button buttonReconnect = mainActivity.findViewById(R.id.reConnectButton);
-        buttonReconnect.setOnClickListener(v -> mainActivity.reconnectToServer());
+        Button buttonReconnect = connectToServerActivity.findViewById(R.id.reConnectButton);
+        buttonReconnect.setOnClickListener(v -> connectToServerActivity.reconnectToServer());
 
     }
